requirements.txt
fastapi==0.116.1
openai==1.101.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
requests==2.32.5
uvicorn[standard]==0.35.0
yt-dlp>=2026.2.4
